requirements.txt
django
whitenoise
waitress
python-dotenv
google-auth
google-auth-httplib2
google-api-python-client
openpyxl